linting and corrections Added CMakeLists.txt Added .gitignore file
